chore: refactor out types and codec for products api
